+++ Source/SwiftLintFramework/Models/MasterRuleList.swift
@@ -13,6 +13,7 @@ public let masterRuleList = RuleList(rules: [
     BlockBasedKVORule.self,
     ClassDelegateProtocolRule.self,
     ClosingBraceRule.self,
+    ClosureBodyLengthRule.self,
     ClosureEndIndentationRule.self,
     ClosureParameterPositionRule.self,
     ClosureSpacingRule.self,